requirements.txt
psycopg2-binary==2.9.7
pandas==2.1.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
schedule==1.2.0
logging==0.4.9.6
configparser==6.0.0
pytest==7.4.0
pytest-cov==4.1.0

